+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 
 ### Misc
 
+- Improvement: Background and Selected Background views for rows and section headers will not be created unless requested. This reduces the overall number of views being allocated.
+
 # Past Releases
 
 # [8.1.0] - 2023-03-16